feat: Fix landing pages behavior
