+++ packages/core/src/components/templates/LandingPage/LandingPage.tsx
@@ -1,38 +1,40 @@
 import { NextSeo, SiteLinksSearchBoxJsonLd } from 'next-seo'
 import type { ComponentType } from 'react'
-import type { GetStaticPaths, GetStaticProps } from 'next'
 import type { Locator } from '@vtex/client-cms'
 
 import RenderSections from 'src/components/cms/RenderSections'
 import Hero from 'src/components/sections/Hero'
-import { mark } from 'src/sdk/tests/mark'
+import Incentives from 'src/components/sections/Incentives'
+import ProductShelf from 'src/components/sections/ProductShelf'
+import ProductTiles from 'src/components/sections/ProductTiles'
+import BannerText from 'src/components/sections/BannerText'
+import Newsletter from 'src/components/sections/Newsletter'
 import { getPage } from 'src/server/cms'
 import type { PageContentType } from 'src/server/cms'
 import CUSTOM_COMPONENTS from 'src/customizations/components'
 
-import storeConfig from '../../../faststore.config'
-import GlobalSections, {
-  GlobalSectionsData,
-  getGlobalSectionsData,
-} from 'src/components/cms/GlobalSections'
+import storeConfig from '../../../../faststore.config'
 
-/**
- * Sections: Components imported from each store's custom components and '../components/sections'.
- * Do not import or render components from any other folder in here.
- */
+/* A list of components that can be used in the CMS. */
 const COMPONENTS: Record<string, ComponentType<any>> = {
   Hero,
+  Incentives,
+  ProductShelf,
+  ProductTiles,
+  BannerText,
+  Newsletter,
   ...CUSTOM_COMPONENTS,
 }
 
-type Props = {
+export type LandingPageProps = {
   page: PageContentType
-  globalSections: GlobalSectionsData
 }
 
-function Page({ page: { sections, settings }, globalSections }: Props) {
+export default function LandingPage({
+  page: { sections, settings },
+}: LandingPageProps) {
   return (
-    <GlobalSections {...globalSections}>
+    <>
       {/* SEO */}
       <NextSeo
         title={settings?.seo?.title ?? storeConfig.seo.title}
@@ -69,46 +71,27 @@ function Page({ page: { sections, settings }, globalSections }: Props) {
         (not the HTML tag) before rendering it here.
       */}
       <RenderSections sections={sections} components={COMPONENTS} />
-    </GlobalSections>
+    </>
   )
 }
 
-export const getStaticProps: GetStaticProps<
-  Props,
-  Record<string, string>,
-  Locator
-> = async ({ previewData, ...context }) => {
-  const [page, globalSections] = await Promise.all([
-    getPage<PageContentType>({
-      ...(previewData?.contentType === 'page'
+export const getLandingPageBySlug = async (
+  slug: string,
+  previewData: Locator
+) => {
+  try {
+    const landingPageData = await getPage<PageContentType>({
+      ...(previewData?.contentType === 'landingPage'
         ? previewData
         : {
             filters: {
-              filters: { 'settings.seo.slug': `/${context.params?.page}` },
+              filters: { 'settings.seo.slug': `/${slug}` },
             },
           }),
-      contentType: 'page',
-    }),
-    getGlobalSectionsData(previewData),
-  ])
-
-  if (!page) {
-    return {
-      notFound: true,
-    }
-  }
-
-  return {
-    props: { page, globalSections },
+      contentType: 'landingPage',
+    })
+    return landingPageData
+  } catch (error) {
+    return null
   }
 }
-
-export const getStaticPaths: GetStaticPaths = async () => {
-  return {
-    paths: [],
-    fallback: 'blocking',
-  }
-}
-
-Page.displayName = 'Page'
-export default mark(Page)
