+++ packages/core/src/components/templates/LandingPage/index.ts
@@ -1,0 +1,2 @@
+export { default, getLandingPageBySlug } from './LandingPage'
+export type { LandingPageProps } from './LandingPage'
