+++ packages/core/src/components/templates/ProductListingPage/index.ts
@@ -1,0 +1,2 @@
+export { default } from './ProductListingPage'
+export type { ProductListingPageProps } from './ProductListingPage'
