+++ packages/core/src/pages/[...slug].tsx
@@ -1,152 +1,51 @@
 import { isNotFoundError } from '@faststore/api'
 import { gql } from '@faststore/graphql-utils'
-import type { SearchState } from '@faststore/sdk'
-import {
-  formatSearchState,
-  parseSearchState,
-  SearchProvider,
-} from '@faststore/sdk'
 import type { GetStaticPaths, GetStaticProps } from 'next'
-import { BreadcrumbJsonLd, NextSeo } from 'next-seo'
-import { useRouter } from 'next/router'
-import { useMemo } from 'react'
 
 import type {
   ServerCollectionPageQueryQuery,
   ServerCollectionPageQueryQueryVariables,
 } from '@generated/graphql'
-import Breadcrumb from 'src/components/sections/Breadcrumb'
-import Hero from 'src/components/sections/Hero'
-import ProductGallery from 'src/components/sections/ProductGallery'
-import ProductShelf from 'src/components/sections/ProductShelf'
-import ScrollToTopButton from 'src/components/sections/ScrollToTopButton'
-import { ITEMS_PER_PAGE, ITEMS_PER_SECTION } from 'src/constants'
-import { useApplySearchState } from 'src/sdk/search/state'
 import { mark } from 'src/sdk/tests/mark'
 import { execute } from 'src/server'
 
 import { Locator } from '@vtex/client-cms'
-import type { ComponentType } from 'react'
 import GlobalSections, {
   getGlobalSectionsData,
   GlobalSectionsData,
 } from 'src/components/cms/GlobalSections'
-import RenderSections from 'src/components/cms/RenderSections'
-import CUSTOM_COMPONENTS from 'src/customizations/components'
-import { getPage, PLPContentType } from 'src/server/cms'
-import storeConfig from '../../faststore.config'
+import { getPage, PageContentType, PLPContentType } from 'src/server/cms'
+import ProductListingPage, {
+  ProductListingPageProps,
+} from 'src/components/templates/ProductListingPage'
+import LandingPage, {
+  getLandingPageBySlug,
+  LandingPageProps,
+} from 'src/components/templates/LandingPage'
 
-type Props = ServerCollectionPageQueryQuery & {
-  page: PLPContentType
+type BaseProps = {
   globalSections: GlobalSectionsData
 }
 
-/**
- * Sections: Components imported from each store's custom components and '../components/sections' only.
- * Do not import or render components from any other folder in here.
- */
-const COMPONENTS: Record<string, ComponentType<any>> = {
-  Breadcrumb,
-  Hero,
-  ProductGallery,
-  ProductShelf,
-  ...CUSTOM_COMPONENTS,
-}
+type Props = BaseProps &
+  (
+    | ({
+        type: 'plp'
+        page: PLPContentType
+      } & ServerCollectionPageQueryQuery)
+    | {
+        type: 'page'
+        page: PageContentType
+      }
+  )
 
-type UseSearchParams = ServerCollectionPageQueryQuery & {
-  sort: SearchState['sort']
-}
-const useSearchParams = ({
-  collection,
-  sort,
-}: UseSearchParams): SearchState => {
-  const selectedFacets = collection?.meta.selectedFacets
-  const { asPath } = useRouter()
-
-  const hrefState = useMemo(() => {
-    const url = new URL(asPath, 'http://localhost')
-
-    const shouldUpdateDefaultSort = sort && !url.searchParams.has('sort')
-    if (shouldUpdateDefaultSort) {
-      url.searchParams.set('sort', sort)
-    }
-
-    const newState = parseSearchState(url)
-    // In case we are in an incomplete url
-    if (newState.selectedFacets.length === 0) {
-      newState.selectedFacets = selectedFacets
-    }
-
-    return formatSearchState(newState).href
-  }, [asPath, selectedFacets, sort])
-
-  return useMemo(() => parseSearchState(new URL(hrefState)), [hrefState])
-}
-
-function Page({
-  page: { sections, settings },
-  globalSections,
-  ...otherProps
-}: Props) {
-  const { collection } = otherProps
-  const router = useRouter()
-  const applySearchState = useApplySearchState()
-  const searchParams = useSearchParams({
-    ...otherProps,
-    sort: settings?.productGallery?.sortBySelection as SearchState['sort'],
-  })
-
-  const { page, sort } = searchParams
-  const title = collection?.seo.title ?? storeConfig.seo.title
-  const description = collection?.seo.description ?? storeConfig.seo.title
-  const pageQuery = page !== 0 ? `?page=${page}` : ''
-  const separator = pageQuery !== '' ? '&' : '?'
-  const sortQuery = !!sort ? `${separator}sort=${sort}` : ''
-  const [pathname] = router.asPath.split('?')
-  const canonical = `${storeConfig.storeUrl}${pathname}${pageQuery}${sortQuery}`
-
+function Page({ globalSections, type, ...otherProps }: Props) {
   return (
     <GlobalSections {...globalSections}>
-      <SearchProvider
-        onChange={applySearchState}
-        itemsPerPage={settings?.productGallery?.itemsPerPage ?? ITEMS_PER_PAGE}
-        {...searchParams}
-      >
-        {/* SEO */}
-        <NextSeo
-          title={title}
-          description={description}
-          titleTemplate={storeConfig.seo.titleTemplate}
-          canonical={canonical}
-          openGraph={{
-            type: 'website',
-            title,
-            description,
-          }}
-        />
-        <BreadcrumbJsonLd
-          itemListElements={collection?.breadcrumbList.itemListElement ?? []}
-        />
-
-        {/*
-          WARNING: Do not import or render components from any
-          other folder than '../components/sections' in here.
-
-          This is necessary to keep the integration with the CMS
-          easy and consistent, enabling the change and reorder
-          of elements on this page.
-
-          If needed, wrap your component in a <Section /> component
-          (not the HTML tag) before rendering it here.
-        */}
-        <RenderSections
-          context={collection}
-          sections={sections}
-          components={COMPONENTS}
-        />
-
-        <ScrollToTopButton />
-      </SearchProvider>
+      {type === 'plp' && (
+        <ProductListingPage {...(otherProps as ProductListingPageProps)} />
+      )}
+      {type === 'page' && <LandingPage {...(otherProps as LandingPageProps)} />}
     </GlobalSections>
   )
 }
@@ -180,7 +79,22 @@ export const getStaticProps: GetStaticProps<
   { slug: string[] },
   Locator
 > = async ({ params, previewData }) => {
-  const [{ data, errors = [] }, page, globalSections] = await Promise.all([
+  const [landingPagePromise, globalSectionsPromise] = [
+    getLandingPageBySlug(params?.slug[0], previewData),
+    getGlobalSectionsData(previewData),
+  ]
+
+  if (await landingPagePromise) {
+    return {
+      props: {
+        page: await landingPagePromise,
+        globalSections: await globalSectionsPromise,
+        type: 'page',
+      },
+    }
+  }
+
+  const [{ data, errors = [] }, page] = await Promise.all([
     execute<
       ServerCollectionPageQueryQueryVariables,
       ServerCollectionPageQueryQuery
@@ -192,7 +106,6 @@ export const getStaticProps: GetStaticProps<
       ...(previewData?.contentType === 'plp' ? previewData : null),
       contentType: 'plp',
     }),
-    getGlobalSectionsData(previewData),
   ])
 
   const notFound = errors.find(isNotFoundError)
@@ -204,6 +117,7 @@ export const getStaticProps: GetStaticProps<
   }
 
   if (errors.length > 0) {
+    console.error(...errors)
     throw errors[0]
   }
 
@@ -211,7 +125,8 @@ export const getStaticProps: GetStaticProps<
     props: {
       ...data,
       page,
-      globalSections,
+      globalSections: await globalSectionsPromise,
+      type: 'plp',
     },
   }
 }
